+++ index.html
@@ -229,10 +229,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-<<<<<<< HEAD
-  <script src="js/main.js"></script>
-=======
-<script src="js/main.js"></script>
->>>>>>> ba509784de32be580854a07547f69cac2fad16a5
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
